Add MCPC+ support for registering entity
This will make the plugin work with MCPC+ servers by inspecting the other potential names the desired method could have.
In order to assist with the behavior of stopping on success, the registration is extracted to a function.

Note that method.setAccessible(true) is not called in "Production environment" - this is because Forge makes the method public.