Rewrite payment code
The main class is now called PaymentManager, and it automatically picks the first PaymentHandler that is enabled and available.
The dt.nocost permissions were redone to make it easier for the PaymentHandlers.
A message was added for when the trip is free (does not apply under FreePaymentHandler).

Some assorted changes were thrown in by accident - the != -> == in FlightsDB was a clear bug.