Implemented command to remove nearest enderdragon
Command "/dt rembugdragon" removes the nearest enderdragon within a
radius of 10 blocks around the player and can be used to remove
stationary dragons which are not  somehow recognized by DragonTravel and
can therefore not be removed using "/dt remstatdragon".
This must not be used on enderdragons recognized by DragonTravel. Those
dragon can be removed by the command as well, but are respawned on next
server startup.